requirements.txt
pandas==1.1.2
requests==2.22.0
textblob==0.15.3
numpy==1.17.4
streamlit==0.68.1
newsapi==0.1.1
PyYAML==5.3.1
